tools/jira_report/requirements.txt
jira
jinja2
pyyaml
